#180 #112 refactor: uuid 추가 및 foodList api refactoring
