+++ src/main/java/com/example/icebutler_server/fridge/dto/fridge/response/RecipeFridgeFoodListRes.java
@@ -8,11 +8,12 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.time.format.DateTimeFormatter;
+import java.util.UUID;
 
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
 public class RecipeFridgeFoodListRes {
-  private Long foodIdx;
+  private UUID foodIdx;
 }
