+++ src/main/java/com/example/icebutler_server/fridge/service/FridgeServiceImpl.java
@@ -297,17 +297,13 @@ public class FridgeServiceImpl implements FridgeService {
   }
 
   //  사용자가 속한 가정용/공용 냉장고 food list
-  public RecipeFridgeFoodListsRes getFridgeUserFoodList(Long fridgeIdx, Long multiFridgeIdx, Long userIdx) {
+  public RecipeFridgeFoodListsRes getFridgeUserFoodList(Long fridgeIdx, Long userIdx) {
     User user = userRepository.findByUserIdxAndIsEnable(userIdx, true).orElseThrow(UserNotFoundException::new);
 
-    if(fridgeIdx != null && multiFridgeIdx == null){
+
       Fridge fridge = this.fridgeRepository.findByFridgeIdxAndIsEnable(fridgeIdx,true).orElseThrow(FridgeNotFoundException::new);
       this.fridgeUserRepository.findByFridgeAndUserAndIsEnable(fridge, user, true).orElseThrow(FridgeUserNotFoundException::new);
       return RecipeFridgeFoodListsRes.toDto(this.fridgeFoodRepository.findByUserForFridgeRecipeFoodList(fridge));
-    }else if(multiFridgeIdx != null && fridgeIdx == null) {
-      MultiFridge fridge = this.multiFridgeRepository.findByMultiFridgeIdxAndIsEnable(multiFridgeIdx, true).orElseThrow(FridgeNotFoundException::new);
-      this.multiFridgeUserRepository.findByMultiFridgeAndUserAndIsEnable(fridge, user, true).orElseThrow(FridgeUserNotFoundException::new);
-      return RecipeFridgeFoodListsRes.toDto(this.fridgeFoodRepository.findByUserForMultiFridgeRecipeFoodList(fridge));
-    }else return new RecipeFridgeFoodListsRes();
+
   }
 }
